refactored more tests
