+++ pg_analytics/tests/insert.rs
@@ -7,13 +7,13 @@ use sqlx::PgConnection;
 
 #[rstest]
 fn insert_user_session_logs(mut conn: PgConnection) {
-    UserSessionLogsTable::setup().execute(&mut conn);
+    UserSessionLogsTable::setup_parquet().execute(&mut conn);
 
     r#"
-    INSERT INTO user_session_logs
-    (event_date, user_id, event_name, session_duration, page_views, revenue)
-    VALUES
-    ('2024-02-01', 2, 'Login', 200, 4, 25.00);
+        INSERT INTO user_session_logs
+        (event_date, user_id, event_name, session_duration, page_views, revenue)
+        VALUES
+        ('2024-02-01', 2, 'Login', 200, 4, 25.00);
     "#
     .execute(&mut conn);
 
@@ -25,35 +25,39 @@ fn insert_user_session_logs(mut conn: PgConnection) {
 
 #[rstest]
 fn insert_user_session_logs_with_null(mut conn: PgConnection) {
-    UserSessionLogsTable::setup().execute(&mut conn);
+    UserSessionLogsTable::setup_parquet().execute(&mut conn);
 
     r#"
-    INSERT INTO user_session_logs
-    (event_date, user_id, event_name, session_duration, page_views, revenue)
-    VALUES
-    (null, null, null, null, null, null);
+        INSERT INTO user_session_logs
+        (event_date, user_id, event_name, session_duration, page_views, revenue)
+        VALUES
+        (null, null, null, null, null, null);
     "#
     .execute(&mut conn);
 
-    let rows: UserSessionLogsRows =
+    let rows: Vec<UserSessionLogsTable> =
         "SELECT * FROM user_session_logs WHERE event_date IS NULL".fetch(&mut conn);
 
+    let first = UserSessionLogsTable {
+        id: 21,
+        event_date: None,
+        user_id: None,
+        event_name: None,
+        session_duration: None,
+        page_views: None,
+        revenue: None,
+    };
+
     assert_eq!(rows.len(), 1);
-    assert_eq!(rows[0].0, Some(21));
-    assert_eq!(rows[0].1, None);
-    assert_eq!(rows[0].2, None);
-    assert_eq!(rows[0].3, None);
-    assert_eq!(rows[0].4, None);
-    assert_eq!(rows[0].5, None);
-    assert_eq!(rows[0].6, None);
+    assert_eq!(rows[0], first);
 }
 
 #[rstest]
 fn insert_research_project_arrays_with_null(mut conn: PgConnection) {
-    ResearchProjectArraysTable::setup().execute(&mut conn);
+    ResearchProjectArraysTable::setup_parquet().execute(&mut conn);
 
     r#"
-    INSERT INTO research_project_arrays (experiment_flags) VALUES (NULL);
+        INSERT INTO research_project_arrays (experiment_flags) VALUES (NULL);
     "#
     .execute(&mut conn);
 
@@ -81,4 +85,102 @@ fn insert_research_project_arrays_with_null(mut conn: PgConnection) {
     };
 
     assert_eq!(rows[0], first);
+}
+
+#[rstest]
+fn insert_not_null(mut conn: PgConnection) {
+    "CREATE TABLE t (a int, b text NOT NULL) USING parquet".execute(&mut conn);
+    "INSERT INTO t VALUES (1, 'test');".execute(&mut conn);
+
+    let row: (i32, String) = "SELECT * FROM t".fetch_one(&mut conn);
+    assert_eq!(row, (1, "test".into()));
+
+    match "INSERT INTO t VALUES (1)".fetch_result::<()>(&mut conn) {
+        Ok(_) => panic!("should not be able to insert null into non-nullable column"),
+        Err(err) => assert_eq!(err.to_string(), "error returned from database: null value in column \"b\" of relation \"t\" violates not-null constraint"),
+    };
+}
+
+#[rstest]
+fn insert_from_series(mut conn: PgConnection) {
+    r#"
+        CREATE TABLE t (
+            id INT
+        ) USING parquet;
+        INSERT INTO t (id) SELECT generate_series(1, 100000);
+        INSERT INTO t (id) SELECT generate_series(1, 100000);
+    "#
+    .execute(&mut conn);
+
+    let count: (i64,) = "SELECT COUNT(*) FROM t".fetch_one(&mut conn);
+    assert_eq!(count, (200000,));
+
+    r#"
+        CREATE TABLE s (
+            id INT
+        ) USING parquet;
+        INSERT INTO s (id) SELECT generate_series(1, 100000);
+        DELETE FROM s WHERE id <= 50000;
+        INSERT INTO s (id) SELECT generate_series(1, 100000);
+    "#
+    .execute(&mut conn);
+
+    let count: (i64,) = "SELECT COUNT(*) FROM s".fetch_one(&mut conn);
+    assert_eq!(count, (150000,));
+}
+
+#[rstest]
+fn insert_parquet_from_parquet(mut conn: PgConnection) {
+    UserSessionLogsTable::setup_parquet().execute(&mut conn);
+    r#"
+        CREATE TABLE copy (
+            id SERIAL PRIMARY KEY,
+            event_date DATE,
+            user_id INT,
+            event_name VARCHAR(50),
+            session_duration INT,
+            page_views INT,
+            revenue DECIMAL(10, 2)
+        ) USING parquet;
+        INSERT INTO copy SELECT * FROM user_session_logs;
+    "#
+    .execute(&mut conn);
+
+    let rows: Vec<(i32, String)> = "SELECT id, event_name FROM copy".fetch(&mut conn);
+
+    let ids = [1, 2, 3, 4, 5, 6, 7, 8, 9, 10];
+    let event_names =
+        "Login,Purchase,Logout,Signup,ViewProduct,AddToCart,RemoveFromCart,Checkout,Payment,Review"
+            .split(',');
+
+    assert!(rows.iter().take(10).map(|r| r.0).eq(ids));
+    assert!(rows.iter().take(10).map(|r| r.1.clone()).eq(event_names));
+}
+
+#[rstest]
+fn insert_parquet_from_heap(mut conn: PgConnection) {
+    UserSessionLogsTable::setup_heap().execute(&mut conn);
+    r#"
+        CREATE TABLE copy (
+            id SERIAL PRIMARY KEY,
+            event_date DATE,
+            user_id INT,
+            event_name VARCHAR(50),
+            session_duration INT,
+            page_views INT,
+            revenue DECIMAL(10, 2)
+        ) USING parquet;
+        INSERT INTO copy SELECT * FROM user_session_logs;
+    "#
+    .execute(&mut conn);
+
+    let rows: Vec<(i32, String)> = "SELECT id, event_name FROM copy".fetch(&mut conn);
+
+    let ids = [1, 2, 3, 4, 5, 6, 7, 8, 9, 10];
+    let event_names =
+        "Login,Purchase,Logout,Signup,ViewProduct,AddToCart,RemoveFromCart,Checkout,Payment,Review"
+            .split(',');
+
+    assert!(rows.iter().take(10).map(|r| r.0).eq(ids));
+    assert!(rows.iter().take(10).map(|r| r.1.clone()).eq(event_names));
 }
